feat/redesign - add new blog cards
